Implement freezing values
Just repeatedly write to the address very quickly.

Also took the time to change how header files were being detected. Now
it fails with an error message if a required header is not found. Same
for required functions.

Additionally, when the usleep function is found it will define the value
for proctal_xopen_flags. This will need to be revised again in the
future. Until then, it just works.