Attempt implementing asynchronous code execution
There are just too many issues with this approach.

There's no way to deallocate the memory where the code is inserted after
it returns execution back to the program's code. We're essentially
causing a memory leak.

Even returning to the original code has proven to be difficult.
Relative jumping has a short address range and the existence of the
x86-64 red zone could corrupt the contents of the stack if I tried to
push an address to trick the ret instruction. Clobbering a register to
do an absolute jump is also not acceptable.

I am going to implement synchronous code execution instead.