Rename function prefixes in the cli source code
The function names were becoming long and difficult to read. Also it was
not a good idea to use the same prefix as the library code, as that made
it difficult to guess whether a function belonged to the library or the
cli.

I could drop the prefix altogether since cli source code is not meant to
be linked to other programs but it makes it easy to do a simple search
and replace for a more unique prefix in case a name clash does occur.