Working fix for instruction pointer backing out
It seems that the previous commit didn't really fix the issue. It might
have to do with alignment requirements. I'll investigate some other
time.

In the mean time this problem will be botched away by placing a bunch of
NOP instructions before the actual code and have the instruction pointer
set in the middle of them so that even if it backs out it will safely
land on some NOP and keep executing the remaining ones until it reaches
the actual code.