Make cli_val_text_parse reject NUL character again
This sort of reverts the changes made in commit
ceb854b30d0b819584fc0c5ba19632472db9f282.

The NUL character is used to mark the end of a C-style string and is
usually not recognized as being part of the text content thus it would
be difficult to distinguish between NUL characters mixed in text and the
end marker that could potentially lead to reading contents past it if
one is not careful enough to check that the function has parsed up to
the end.

It would make sense to add a binary safe alternative for parsing text.

Also added more test suites and adjusted existing ones.