Introduce the init script
This script is used to fetch dependencies and prepare the autotools.
Some of this work used to be done by the configure script but it made
the release builds unnecessarily more difficult for the user by
requiring the autotools installed which defeats the purpose of
generating the configure script in the first place.