Remove INSTALL and COPYING files
These were created when I was first starting off using the autotools for
this project and let it generate the initial structure.

The generic INSTALL file seems to be too tedious for a user to digest.
The README file and the help information produced by the configure
script seem good enough. Anything else specific to the autotools can be
looked up online.

Now distributing just a LICENSE file.