Drop `proctal_global_` prefix
The idea was to have all functions that do not operate on a handle get
the prefix `proctal_global_` but it just doesn't look aesthetically
pleasing. To be consistent I would have to rename the function
`proctal_open` to `proctal_global_open` while `proctal_close` would
continue to have the same name and thus no longer look symmetrical.