Use const keyword for immutable struct darr arrays
The const keyword will clearly indicate that those structures are not
meant to be modified.

The functions proctal_global_free and proctal_free now take a const void
pointer to support these kinds of constructs.

The darr library was updated to version 1.7.0 that comes with support
for the const keyword.