Successfully cross-compile to MinGW-w64 from Linux
Most features remain unimplemented but what's important is that it
actually compiles now.

Some functions from the C standard library implemented by MinGW-w64 are
declared as static functions in headers. This makes it impossible to use
functions such as printf in inline functions that are not static. Hence
why some functions are no longer marked inline.

While Windows has a POSIX subsystem, it lacks a lot of features. This
particularly hinders what can be done with streams and signals. The
--input option from the freeze command had to be removed to make porting
easier. The signal handling code still needs to be rewritten for
Windows.

Had to place return statements that are never executed so that the
compiler does not complain that the functions end without returning a
value.