Address for watch command is passed as an argument
With the introduction of the address range options (--addresss-start and
--address-stop) that share similar names to the --address option there
was a possibility that users could get confused by their behavor. While
the --address option specified what address to watch, --address-start
and --address-stop narrowed the locations in memory that would get
reported for accessing the address.

By removing the option and having the user pass the address as an
argument we can avoid that confusion and additionally make the command
look simpler.

Also took the opportunity to fix descriptions, grammar and example code.