Replaced instruction value type
It's not going to be feasible to add options to the instruction type
that only apply to certain architectures. The --instruction-syntax
option only made sense to x86, for instance. The instruction type will
become more complicated to document the more architectures it supports.

With the instruction value type not being a good abstraction, it has
been split into 3 new types:

- x86
- arm
- aarch64

This will allow each architecture to be configured independently from
each other.

x86 now supports 16-bit mode.